+++ AccountsReceivable.API/Models/CustomerWalletInfo.cs
@@ -9,16 +9,16 @@ namespace AccountsReceivable.API.Models
         [Key]
         public int CustomerWalletId { get; set; }
         public int CustomerId { get; set; }
-        public int? TotalBusinessAmount { get; set; }
-        public int? TotalPaidAmount { get; set; }
-        public int? DueAmount { get; set; }
-        public int? CreditLimit { get; set; }
+        public decimal? TotalBusinessAmount { get; set; }
+        public decimal? TotalPaidAmount { get; set; }
+        public decimal? DueAmount { get; set; }
+        public decimal? CreditLimit { get; set; }
     }
     public class CustomerWalletTransactionList
     {
         [Key]
         public int? TransactionId { get; set; }
-        public int? TransactionAmount { get; set; }
+        public decimal? TransactionAmount { get; set; }
         public string TransactionType { get; set; }
         public string TransactionMode { get; set; }
         public string CardNumber { get; set; }
