+++ AccountsReceivable.API/Models/RequestModel/CashbackMasterRequest.cs
@@ -5,11 +5,11 @@ namespace AccountsReceivable.API.Models.RequestModel
     {
         //public int CashBackMasterId { get; set; }
         public int? CompanyId { get; set; }
-        public int? MinimumBusinessAmount { get; set; }
-        public int? MaximumCashbackAmount { get; set; }
+        public decimal? MinimumBusinessAmount { get; set; }
+        public decimal? MaximumCashbackAmount { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int? CashbackValue { get; set; }
+        public decimal? CashbackValue { get; set; }
         public int? IsPercentage { get; set; }
         public Boolean IsActive { get; set; }
